Refactor: Further split UVIslandMaskEditor into partial classes
Extracted two more major regions to separate partial class files:
- Mouse Event Handlers (305 lines) → UVIslandMaskEditor.Events.cs
- Async Initialization (389 lines) → UVIslandMaskEditor.Async.cs

File size improvements:
- Main file: 2991 → 1447 lines (-51.6% reduction)
- All files now under 1500 lines (ideal for maintainability)

File structure:
├─ UVIslandMaskEditor.cs (main logic, 1447 lines)
├─ UVIslandMaskEditor.UI.cs (UI creation, 871 lines)
├─ UVIslandMaskEditor.Events.cs (mouse events, 325 lines)
└─ UVIslandMaskEditor.Async.cs (async init, 409 lines)

This completes the file size optimization phase. The codebase is
now ready for lifecycle management simplification.
