+++ ExDeform/Editor/UVIslandMaskEditor.cs
@@ -433,312 +433,6 @@ namespace DeformEditor.Masking
 
         #endregion
 
-        #region Mouse Event Handlers
-        // マウスイベントハンドラ
-        // Mouse event handling for UV map interaction
-
-        private void OnUVMapMouseDown(MouseDownEvent evt)
-        {
-            if (selector == null) return;
-            
-            // Any mouse interaction should trigger full resolution mode
-            OnUserInteraction();
-            
-            var localPosition = evt.localMousePosition;
-            
-            if (evt.button == 0) // Left click
-            {
-                if (isMagnifyingGlassActive)
-                {
-                    evt.StopPropagation();
-                    return;
-                }
-                
-                if (selector.EnableRangeSelection && evt.shiftKey)
-                {
-                    // Check for deselection mode (Ctrl+Shift) at the start of range selection
-                    isRangeDeselecting = evt.ctrlKey && evt.shiftKey;
-                    StartRangeSelection(localPosition);
-                }
-                else
-                {
-                    HandleIslandSelection(localPosition);
-                }
-                evt.StopPropagation();
-            }
-            else if (evt.button == 2 && !isMagnifyingGlassActive) // Middle click for pan
-            {
-                isDraggingUVMap = true;
-                lastMousePos = localPosition;
-                evt.StopPropagation();
-            }
-            else if (evt.button == 1 && selector.EnableMagnifyingGlass) // Right click for magnifying glass
-            {
-                StartMagnifyingGlass(localPosition);
-                evt.StopPropagation();
-                return;
-            }
-        }
-        
-        private void HandleIslandSelection(Vector2 localPosition)
-        {
-            // Use proper coordinate transformation that accounts for zoom and pan
-            var uvCoordinate = LocalPosToUV(localPosition);
-            
-            int islandID = selector.GetIslandAtUVCoordinate(uvCoordinate);
-            
-            if (islandID >= 0)
-            {
-                // User interaction detected - switch to full resolution mode
-                OnUserInteraction();
-
-                Undo.RecordObject(targetMask, "Toggle UV Island Selection");
-                selector.ToggleIslandSelection(islandID);
-                UpdateMaskComponent();
-                EditorUtility.SetDirty(targetMask);
-
-                // Generate full texture and update display
-                if (selector.AutoUpdatePreview)
-                {
-                    selector.GenerateUVMapTexture();
-                    RefreshUVMapImage();
-                }
-
-                // Save low-res texture to cache after selection changes
-                SaveLowResTextureToCache();
-                
-                RefreshUIFast();
-            }
-            // Removed automatic pan start on left click - pan is now middle button only
-        }
-        
-        private void OnUVMapMouseMove(MouseMoveEvent evt) => HandleMouseMove(evt);
-        private void OnUVMapContainerMouseMove(MouseMoveEvent evt) => HandleMouseMove(evt);
-        
-        private void HandleMouseMove(MouseMoveEvent evt)
-        {
-            if (selector == null) return;
-            
-            var localPosition = evt.localMousePosition;
-            
-            if (isMagnifyingGlassActive)
-            {
-                UpdateMagnifyingGlass(localPosition);
-                evt.StopPropagation();
-            }
-            else if (selector.IsRangeSelecting)
-            {
-                UpdateRangeSelection(localPosition);
-                evt.StopPropagation();
-            }
-            else if (isDraggingUVMap && !isMagnifyingGlassActive)
-            {
-                var deltaPos = localPosition - lastMousePos;
-                // Fixed pan sensitivity to maintain consistent movement regardless of zoom
-                var panSensitivity = 1f / UV_MAP_SIZE;
-                var uvDelta = new Vector2(
-                    deltaPos.x * panSensitivity,
-                    -deltaPos.y * panSensitivity
-                );
-                
-                var currentOffset = selector.UvMapPanOffset;
-                selector.SetPanOffset(currentOffset + uvDelta);
-                
-                lastMousePos = localPosition;
-                
-                if (selector.AutoUpdatePreview)
-                {
-                    UpdateTextureWithThrottle(); // Immediate feedback for pan with throttling
-                }
-                
-                evt.StopPropagation();
-            }
-        }
-        
-        private void OnUVMapMouseUp(MouseUpEvent evt) => HandleMouseUp(evt);
-        private void OnUVMapContainerMouseUp(MouseUpEvent evt) => HandleMouseUp(evt);
-        
-        private void HandleMouseUp(MouseUpEvent evt)
-        {
-            if (evt.button == 0) // Left button
-            {
-                if (isMagnifyingGlassActive)
-                {
-                    HandleMagnifyingGlassClick(evt);
-                }
-                else if (selector?.IsRangeSelecting == true)
-                {
-                    bool addToSelection = evt.shiftKey && !evt.ctrlKey;
-                    bool removeFromSelection = evt.ctrlKey && evt.shiftKey;
-                    isRangeDeselecting = removeFromSelection;
-                    FinishRangeSelection(addToSelection, removeFromSelection);
-                }
-                
-                evt.StopPropagation();
-            }
-            else if (evt.button == 2) // Middle button - stop panning
-            {
-                isDraggingUVMap = false;
-                
-                // Update texture after mouse interaction ends
-                if (selector?.AutoUpdatePreview ?? false)
-                {
-                    selector?.UpdateTextureIfNeeded();
-                    RefreshUVMapImage();
-                }
-                
-                evt.StopPropagation();
-            }
-            else if (evt.button == 1) // Right button
-            {
-                StopMagnifyingGlass();
-                evt.StopPropagation();
-            }
-        }
-        
-        private void OnUVMapWheel(WheelEvent evt)
-        {
-            if (selector == null) return;
-            
-            // Zoom interaction should trigger full resolution mode
-            OnUserInteraction();
-            
-            var localPosition = evt.localMousePosition;
-            var zoomPoint = LocalPosToUV(localPosition);
-            var zoomDelta = -evt.delta.y * 0.1f;
-            
-            selector.ZoomAtPoint(zoomPoint, zoomDelta);
-            zoomSlider.value = selector.UvMapZoom;
-            
-            if (selector.AutoUpdatePreview)
-            {
-                UpdateTextureWithThrottle(); // Immediate feedback for wheel zoom
-            }
-            
-            evt.StopPropagation();
-        }
-        
-        private void OnRootMouseMove(MouseMoveEvent evt)
-        {
-            if (selector == null) return;
-            
-            var containerWorldBound = uvMapContainer.worldBound;
-            var relativeX = evt.mousePosition.x - containerWorldBound.x;
-            var relativeY = evt.mousePosition.y - containerWorldBound.y;
-            var localPos = new Vector2(relativeX, relativeY);
-            
-            if (selector.IsRangeSelecting)
-            {
-                var clampedPos = new Vector2(
-                    Mathf.Clamp(localPos.x, 0, UV_MAP_SIZE),
-                    Mathf.Clamp(localPos.y, 0, UV_MAP_SIZE)
-                );
-                
-                var uvCoord = LocalPosToUV(clampedPos);
-                selector.UpdateRangeSelection(uvCoord);
-                
-	            bool removeFromSelection = evt.ctrlKey && evt.shiftKey;
-                // Update deselection mode state based on current key state during dragging
-                // Use Input class for cross-platform key detection
-	            isRangeDeselecting = removeFromSelection;
-                
-                UpdateRangeSelectionVisual();
-                evt.StopPropagation();
-            }
-            else if (isMagnifyingGlassActive)
-            {
-                var clampedPos = new Vector2(
-                    Mathf.Clamp(localPos.x, 0, UV_MAP_SIZE),
-                    Mathf.Clamp(localPos.y, 0, UV_MAP_SIZE)
-                );
-                
-                UpdateMagnifyingGlass(clampedPos);
-                evt.StopPropagation();
-            }
-            else if (isDraggingUVMap)
-            {
-                var clampedPos = new Vector2(
-                    Mathf.Clamp(localPos.x, 0, UV_MAP_SIZE),
-                    Mathf.Clamp(localPos.y, 0, UV_MAP_SIZE)
-                );
-                
-                var deltaPos = clampedPos - lastMousePos;
-                // Fixed pan sensitivity to maintain consistent movement regardless of zoom
-                var panSensitivity = 1f / UV_MAP_SIZE;
-                var uvDelta = new Vector2(
-                    deltaPos.x * panSensitivity,
-                    -deltaPos.y * panSensitivity
-                );
-                
-                var currentOffset = selector.UvMapPanOffset;
-                selector.SetPanOffset(currentOffset + uvDelta);
-                
-                lastMousePos = clampedPos;
-                
-                if (selector.AutoUpdatePreview)
-                {
-                    UpdateTextureWithThrottle(); // Immediate feedback for pan with throttling
-                }
-                
-                evt.StopPropagation();
-            }
-        }
-        
-        private void OnRootMouseUp(MouseUpEvent evt)
-        {
-            if (selector == null) return;
-            
-            if (evt.button == 0)
-            {
-                if (selector.IsRangeSelecting)
-                {
-                    bool addToSelection = evt.shiftKey && !evt.ctrlKey;
-                    bool removeFromSelection = evt.ctrlKey && evt.shiftKey;
-                    isRangeDeselecting = removeFromSelection;
-                    FinishRangeSelection(addToSelection, removeFromSelection);
-                    evt.StopPropagation();
-                }
-                else if (isDraggingUVMap)
-                {
-                    isDraggingUVMap = false;
-                    
-                    // Update texture after mouse interaction ends
-                    if (selector?.AutoUpdatePreview ?? false)
-                    {
-                        selector?.UpdateTextureIfNeeded();
-                        RefreshUVMapImage();
-                    }
-                    
-                    evt.StopPropagation();
-                }
-            }
-            else if (evt.button == 2)
-            {
-                if (isDraggingUVMap)
-                {
-                    isDraggingUVMap = false;
-                    
-                    // Update texture after mouse interaction ends
-                    if (selector?.AutoUpdatePreview ?? false)
-                    {
-                        selector?.UpdateTextureIfNeeded();
-                        RefreshUVMapImage();
-                    }
-                    
-                    evt.StopPropagation();
-                }
-            }
-            else if (evt.button == 1)
-            {
-                if (isMagnifyingGlassActive)
-                {
-                    StopMagnifyingGlass();
-                    evt.StopPropagation();
-                }
-            }
-        }
-
-        #endregion
 
         #region Mesh and Data Access
         // メッシュとデータアクセス
@@ -963,396 +657,6 @@ namespace DeformEditor.Masking
 
         #endregion
 
-        #region Async Initialization
-        // 非同期初期化
-        // Async initialization and progress monitoring
-
-        /// <summary>
-        /// Show placeholder message while UV map is being initialized
-        /// UV マップ初期化中のプレースホルダーメッセージを表示
-        /// </summary>
-        private void ShowPlaceholderMessage(string message)
-        {
-            // Update status label
-            if (statusLabel != null)
-            {
-                statusLabel.text = message;
-            }
-
-            // Show placeholder in UV map area
-            if (uvMapImage != null)
-            {
-                // Set a neutral gray background to indicate loading state
-                uvMapImage.style.backgroundColor = new StyleColor(new Color(0.22f, 0.22f, 0.22f, 1f));
-                // Image will be replaced when texture loads, no need to explicitly clear
-            }
-
-            // Clear island list during initialization
-            if (islandListView != null)
-            {
-                islandListView.itemsSource = null;
-                islandListView.Rebuild();
-            }
-        }
-
-        // ShowProgressUI and HideProgressUI methods removed - replaced with InitializationProgressView
-
-        /// <summary>
-        /// Monitor async initialization progress and update UI elements
-        /// 非同期初期化の進捗を監視し、UI要素を更新
-        /// </summary>
-        private void MonitorAsyncInitialization()
-        {
-            if (asyncInitManager == null || !asyncInitManager.IsRunning)
-            {
-                // Stop monitoring when initialization is not running
-                EditorApplication.update -= MonitorAsyncInitialization;
-
-                // Hide progress view when monitoring stops
-                if (progressView != null)
-                {
-                    progressView.style.display = DisplayStyle.None;
-                }
-                return;
-            }
-
-            // Update progress view with latest status
-            if (progressView != null)
-            {
-                progressView.Progress = asyncInitManager.Progress;
-                progressView.StatusMessage = asyncInitManager.StatusMessage;
-            }
-
-            // Update status label
-            if (statusLabel != null)
-            {
-                statusLabel.text = asyncInitManager.StatusMessage;
-            }
-
-            // Update UV image with incremental textures
-            if (selector != null && selector.UvMapTexture != null && uvMapImage != null)
-            {
-                uvMapImage.style.backgroundImage = new StyleBackground(selector.UvMapTexture);
-            }
-        }
-
-        /// <summary>
-        /// Callback when async initialization is completed
-        /// 非同期初期化完了時のコールバック
-        /// </summary>
-        private void OnAsyncInitializationCompleted(UVIslandSelector completedSelector)
-        {
-            Debug.Log("[UVIslandMaskEditor] OnAsyncInitializationCompleted called");
-
-            // Stop monitoring async initialization
-            EditorApplication.update -= MonitorAsyncInitialization;
-
-            // Clear async initialization flag
-            asyncInitializationInProgress = false;
-
-            if (completedSelector == null)
-            {
-                Debug.LogError("[UVIslandMaskEditor] Completion callback received null selector");
-                if (progressView != null)
-                {
-                    progressView.style.display = DisplayStyle.None;
-                }
-                if (statusLabel != null)
-                {
-                    statusLabel.text = "Initialization failed";
-                }
-                return;
-            }
-
-            Debug.Log($"[UVIslandMaskEditor] Selector initialized with {completedSelector.UVIslands?.Count ?? 0} islands");
-
-            // Step 1: Restore island selections now that UV analysis is complete
-            if (targetMask != null)
-            {
-                // Load per-submesh selections (new format)
-                if (targetMask.PerSubmeshSelections.Count > 0)
-                {
-                    Debug.Log($"[UVIslandMaskEditor] Restoring {targetMask.PerSubmeshSelections.Count} per-submesh selections");
-                    selector.SetAllSelectedIslands(targetMask.GetPerSubmeshSelections());
-                }
-                // Fallback to legacy flat list if new format is empty (backward compatibility)
-                else if (targetMask.SelectedIslandIDs.Count > 0)
-                {
-                    Debug.Log($"[UVIslandMaskEditor] Restoring {targetMask.SelectedIslandIDs.Count} legacy island selections");
-                    selector.SetSelectedIslands(targetMask.SelectedIslandIDs);
-                }
-            }
-
-            // Step 2: Update flags BEFORE UI refresh
-            textureInitialized = true;
-            isInitialized = true;
-            isLoadingFromCache = false;
-            shouldShowLowResUntilInteraction = false;
-
-            // Step 3: Re-enable auto preview
-            selector.AutoUpdatePreview = true;
-
-            // Step 4: Clear placeholder background color
-            if (uvMapImage != null)
-            {
-                uvMapImage.style.backgroundColor = StyleKeyword.Null;
-                uvMapImage.MarkDirtyRepaint();
-            }
-
-            // Step 5: Save low-res texture to cache for next reload
-            SaveLowResTextureToCache();
-
-            // Step 6: Full UI refresh to update all elements
-            Debug.Log("[UVIslandMaskEditor] Refreshing UI after initialization");
-            RefreshUI(false);
-
-            // Step 7: Update status message
-            if (statusLabel != null)
-            {
-                int islandCount = selector.UVIslands?.Count ?? 0;
-                statusLabel.text = UVIslandLocalization.Get("status_islands_found", islandCount);
-            }
-
-            // Step 8: Hide progress view LAST to ensure all UI is updated first
-            if (progressView != null)
-            {
-                progressView.style.display = DisplayStyle.None;
-            }
-
-            Debug.Log("[UVIslandMaskEditor] Async initialization completed successfully");
-        }
-
-        /// <summary>
-        /// Force immediate data refresh with texture generation - used for initial load
-        /// </summary>
-        private void RefreshDataWithImmediteTexture()
-        {
-            if (selector == null) 
-            {
-                if (statusLabel != null)
-                {
-                    statusLabel.text = "No mesh data available";
-                }
-                return;
-            }
-            
-            if (statusLabel != null)
-            {
-                statusLabel.text = UVIslandLocalization.Get("status_refreshing");
-            }
-            
-            try 
-            {
-                // Force mesh data update
-                selector.UpdateMeshData();
-                
-                // Always generate texture immediately on first load
-                selector.GenerateUVMapTexture();
-                textureInitialized = true;
-                isInitialized = true;
-                isLoadingFromCache = false; // Clear cache loading flag since full texture is ready
-                shouldShowLowResUntilInteraction = false; // Ensure we show full-res texture, not low-res
-
-                // Save low-res texture to cache for next reload
-                SaveLowResTextureToCache();
-
-                // Clear placeholder background color
-                if (uvMapImage != null)
-                {
-                    uvMapImage.style.backgroundColor = StyleKeyword.Null;
-                }
-
-                // Immediate UI refresh
-                RefreshUVMapImage();
-
-                if (selector?.UVIslands != null)
-                {
-                    islandListView.itemsSource = selector.UVIslands;
-                    islandListView.Rebuild(); // Use Rebuild instead of RefreshItems to ensure full update
-                }
-
-                // Update UI elements that were created with null selector
-                UpdateSubmeshSelectorUI();
-                UpdateHighlightSettingsUI();
-                UpdateDisplaySettingsUI();
-                UpdateSubmeshLabel();
-
-                UpdateStatus();
-
-                int islandCount = selector.UVIslands?.Count ?? 0;
-                if (statusLabel != null)
-                {
-                    statusLabel.text = UVIslandLocalization.Get("status_islands_found", islandCount);
-                }
-            }
-            catch (System.Exception ex)
-            {
-                if (statusLabel != null)
-                {
-                    statusLabel.text = $"Error: {ex.Message}";
-                }
-                Debug.LogError($"[UVIslandMaskEditor] Error refreshing data: {ex}");
-            }
-        }
-        
-        private void RefreshUI(bool forceSceneRepaint = false)
-        {
-            // Always refresh the texture when UI updates
-            if (selector?.AutoUpdatePreview ?? false)
-            {
-                selector.UpdateTextureIfNeeded(); // Use deferred update instead of direct generation
-            }
-            
-            RefreshUVMapImage();
-            
-            if (selector?.UVIslands != null)
-            {
-                islandListView.itemsSource = selector.UVIslands;
-                islandListView.RefreshItems();
-            }
-            
-            UpdateStatus();
-            
-            if (forceSceneRepaint)
-            {
-                SceneView.RepaintAll();
-            }
-        }
-        
-        // Fast UI refresh for frequent operations like selection changes
-        private void RefreshUIFast()
-        {
-            // Update essential UI elements immediately
-            UpdateStatus();
-            
-            // Update list view selection state without full refresh
-            if (islandListView != null && selector?.UVIslands != null)
-            {
-                // Only refresh if needed
-                if (islandListView.itemsSource != selector.UVIslands)
-                {
-                    islandListView.itemsSource = selector.UVIslands;
-                }
-                islandListView.RefreshItems();
-            }
-            
-            // Only generate texture if needed and not showing low-res until interaction
-            if (selector != null && !shouldShowLowResUntilInteraction)
-            {
-                if (selector.UvMapTexture == null)
-                {
-                    selector.GenerateUVMapTexture();
-                    textureInitialized = true;
-                }
-            }
-            
-            // Always refresh image (may show low-res or full-res based on state)
-            RefreshUVMapImage();
-
-            // Only repaint scene if there are selected islands to display
-            if (selector?.HasSelectedIslands ?? false)
-            {
-                SceneView.RepaintAll();
-            }
-        }
-        
-        private void RefreshUVMapImage()
-        {
-            if (selector?.UvMapTexture != null && !shouldShowLowResUntilInteraction)
-            {
-                // Show full resolution texture
-                uvMapImage.style.backgroundImage = new StyleBackground(selector.UvMapTexture);
-                ClearLowResDisplayState();
-            }
-            else if (currentLowResTexture != null && (isLoadingFromCache || shouldShowLowResUntilInteraction))
-            {
-                // Show low-resolution cached texture until user interaction
-                uvMapImage.style.backgroundImage = new StyleBackground(currentLowResTexture);
-            }
-            else if (selector?.UvMapTexture != null)
-            {
-                // Fallback to full texture if low-res is not available
-                uvMapImage.style.backgroundImage = new StyleBackground(selector.UvMapTexture);
-                ClearLowResDisplayState();
-            }
-            else
-            {
-                // Clear image if no texture is available
-                uvMapImage.style.backgroundImage = StyleKeyword.None;
-            }
-        }
-        
-        /// <summary>
-        /// Centralized method to clear low-res display state
-        /// </summary>
-        private void ClearLowResDisplayState()
-        {
-            isLoadingFromCache = false;
-            shouldShowLowResUntilInteraction = false;
-        }
-        
-        // Throttled immediate texture update for interactive operations
-        private void UpdateTextureWithThrottle()
-        {
-            if (selector == null) return;
-            
-            float currentTime = Time.realtimeSinceStartup;
-            if (currentTime - lastUpdateTime >= TEXTURE_UPDATE_THROTTLE)
-            {
-                // Immediate update if enough time has passed
-                selector.GenerateUVMapTexture();
-                RefreshUVMapImage();
-                lastUpdateTime = currentTime;
-            }
-            else if (pendingTextureUpdate == null)
-            {
-                // Schedule single deferred update if throttled and none pending
-                pendingTextureUpdate = () =>
-                {
-                    if (selector != null)
-                    {
-                        selector.GenerateUVMapTexture();
-                        RefreshUVMapImage();
-                        lastUpdateTime = Time.realtimeSinceStartup;
-                    }
-                    pendingTextureUpdate = null;
-                };
-                EditorApplication.delayCall += pendingTextureUpdate;
-            }
-            // If there's already a pending update, do nothing to avoid duplicates
-        }
-        
-        private void UpdateStatus()
-        {
-            if (selector == null) return;
-            
-            // Show total selected islands across all submeshes
-            var totalSelectedCount = selector.AllSelectedIslandIDs?.Count ?? 0;
-            var currentSubmeshSelectedCount = selector.SelectedIslandIDs?.Count ?? 0;
-            var maskedVertexCount = selector.VertexMask?.Length ?? 0;
-            var maskedFaceCount = (selector.TriangleMask?.Length ?? 0) / 3;
-
-            if (totalSelectedCount > 0)
-            {
-                // Show both current submesh selection and total across all submeshes
-                if (selector.SelectedSubmeshIndices.Count > 1)
-                {
-                    statusLabel.text = $"Selected: {currentSubmeshSelectedCount} islands (Submesh {selector.CurrentPreviewSubmesh}) | Total: {totalSelectedCount} islands across all submeshes | {maskedVertexCount} vertices, {maskedFaceCount} faces";
-                }
-                else
-                {
-                    statusLabel.text = UVIslandLocalization.Get("status_islands_selected",
-                        currentSubmeshSelectedCount, maskedVertexCount, maskedFaceCount);
-                }
-            }
-            else
-            {
-                int islandCount = selector.UVIslands?.Count ?? 0;
-                statusLabel.text = UVIslandLocalization.Get("status_islands_found", islandCount);
-            }
-        }
-
-        #endregion
 
         #region Island Selection and Interaction
         // アイランド選択とインタラクション
